+++ source/blender/makesdna/DNA_brush_types.h
@@ -59,7 +59,7 @@ typedef struct Brush {
 	int flag;					/* general purpose flag */
 	float detail;			/* dynamic subdivission detail */
 	float jitter;				/* jitter the position of the brush */
-	float spacing;				/* spacing of paint operations */
+	int spacing;				/* spacing of paint operations */
 	int smooth_stroke_radius;		/* turning radius (in pixels) for smooth stroke */
 	float smooth_stroke_factor;		/* higher values limit fast changes in the stroke direction */
 	float rate;					/* paint operations / second (airbrush) */
